Change configuration storage and add Supression file
The previous configuration storage container was a dictionary. Because
of this the same glob pattern (key) cannot be used more than once. This
is solved by introducing a collection of tuples.

The suppression file was added because of a ruleset warning about CLS
compliant code. The MSTest attributes "DataTestMethod" and "DataRow"
make the test project not CLS compliant.