GH-19 Replace file search pattern with glob
I have previously removed glob patterns because I was not happy with how
the functioned. This was on one hand because of a self inflicted bug and
on the other hand because of how some frameworks handle globs. The new
implementation is not as performant as the file search pattern approach,
but I believe that glob patterns are more intuitive and more flexible