GH-20 Redesign Bumpy CLI
Fixes #20

This commit redesigns Bumpy's CLI to be more in lign with other command
line tools such as chocolatey or git. Previous commands such as "bumpy
-l" will now use full words such as "bumpy list". This change made it
possible to add options in a non-confusing way. E.g. specifing a profile
would look like this: "bumpy list -p my_profile".

This commit also adds new options which make it possible to specify a
different directory as well as different location where Bumpy will
search for configuration.

Even though this is a pretty significant breaking change, I strongly
believe that it is the correct way going forward. It would be very easy
to add new options to existing commands.