Remove null validation checks
I am a big fan of avoiding NULL as much as possible, which is why I was
using a `ThrowIfNull` extension method in the first place. As I am not
planning to release the core of Bumpy as a library, any real cases
in which NULL could enter the application via a user by accident are not
existant. Or let's say that I cannot think of any case right now.
Because of this realisation I am removing these checks for now